% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areParsed.R
\name{areTagsParsed}
\alias{areTagsParsed}
\alias{areReadTagsParsed}
\alias{areReadTagsParsed.SamReads}
\alias{areReadTagsParsed.Sam}
\alias{areHeaderTagsParsed}
\alias{areHeaderTagsParsed.SamHeader}
\alias{areHeaderTagsParsed.Sam}
\title{Are sam tags parsed}
\usage{
areReadTagsParsed(x, ...)

\method{areReadTagsParsed}{SamReads}(x, ...)

\method{areReadTagsParsed}{Sam}(x, ...)

areHeaderTagsParsed(x, ...)

\method{areHeaderTagsParsed}{SamHeader}(x, ...)

\method{areHeaderTagsParsed}{Sam}(x, ...)
}
\arguments{
\item{x}{The Sam, SamReads or SamHeader object to check for parsed tags.}

\item{...}{Required for S3 object method implementation. Not currently used.}
}
\value{
TRUE if tags are parsed, FALSE otherwise.
}
\description{
Accessors for checking if the data associated with this Sam file has parsed
tags, or has one column with tags unparsed. Both reads and headers will have
a parsed and an unparsed version, although currently only an unparsed version
is avaialble for headers.
}
