% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samFlags.R
\name{checkSamFlag}
\alias{checkSamFlag}
\title{Do sam values match selected/unselected flags}
\usage{
checkSamFlag(flagVal, isFlag = "", notFlag = "")
}
\arguments{
\item{flagVal}{A vector of sam flag integer values representin the 12 sam
read flags.}

\item{isFlag}{A vector of flag (names) that must be \code{TRUE}, by default
none. Returns \code{TRUE} only if the \code{flagVal} indicates that all of
these flags are set.}

\item{isFlag}{A vector of flag names that must be \code{FALSE}, by default
none.  Returns \code{TRUE} only if the \code{flagVal} indicates that none
of these flags are set}
}
\value{
Returns a logical vector of the same length as \code{flagVal}. TRUE
  if all \code{isFlag} flags are set and no \code{notFlag} flag is set. The
  status of flags listed in neither \code{isFlag} nor \code{notFlag} does not
  matter. Note - if you specify the same flag in both lists, all
  \code{flagVal} are FALSE. If you specify nothing in either, all are TRUE.
  If you specify all flags as either TRUE or FALSE (i.e. no flag is
  ambiguous/unimportant), it would be better to check flagVal ==
  samFlag(isFlag).
}
\description{
Given a vector of sam flag values, return true if the value indicates that
all \code{isFlag} flags are set and that none of the \code{notFlag} flags are
set.
}
\examples{
matchFlags( 1:16, isFlag=  c( "READ_PAIRED", "PROPER_PAIR" ),
                  notFlag= c( "READ_UNMAPPED" ))
#> [1] FALSE FALSE  TRUE FALSE FALSE FALSE FALSE FALSE
#> [9] FALSE FALSE  TRUE FALSE FALSE FALSE FALSE FALSE

}

