% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exceptions.R
\name{condition}
\alias{condition}
\title{Create a condition object.}
\usage{
condition(message = "condition", call = NULL)
}
\arguments{
\item{message}{A one-element character vector describing this
condition/event. By default this is \code{"condition"}. It is fine to set
this NULL as some subclasess will not need it (i.e. an extension class
that only holds named data.)}

\item{call}{The call or calls that were in process when this condition or
event was generated. By default this is \code{NULL} as conditions need not
be code related events (e.g. may be due to events triggered by
environmental conditions or timers, etc.). \code{sys.calls()} can be used
to obtain the whole call stack, and \code{sys.call()} can be used to access
a single call.}
}
\value{
A condition object, of class "condition".
}
\description{
Create an S3 object of class \code{"condition"} for use in \code{R}'s
condition handling system and as a base class for an exception class hierarchy.
}
\examples{
e <- condition()
conditionMessage( e )
conditionCall( e )

cond <- condition( "Something happened.", call= sys.calls() )

\dontrun{
   # Catch a "condition" class error and turn it into a warning instead.
   tryCatch(
        { stop(cond) },
        condition= function(e) { warning(e) }
   )
}
}

