% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamReads.R
\name{countAllReads}
\alias{countAllReads}
\title{Read statistics}
\usage{
countAllReads(x, ...)
}
\arguments{
\item{x}{A \code{Sam} or \code{SamReads} object to return read statistics for.

Returns the number of lines of reads in the Sam file. Does no filtering, so
counts a multi-aligned read end once for each alignment and includes
duplicates, unaligned reads, failed QC alignments, etc. Can filter and then
re-apply to get any sub-scount desired. Dividing by 2 does not convert to
paired end results as some ends may match more or less times than another.}

\item{...}{Required for S3 object method implementation. Not currently used.}
}
\value{
The total number of read records in the Sam file object.
}
\description{
Read statistics
}
