% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataExceptions.R
\name{DataException}
\alias{DataException}
\alias{EmptyFileException}
\alias{FileFormatException}
\alias{FileNotEmptyException}
\title{The DataException hierarchy}
\usage{
DataException(data = NA, message = "A DataException occurred.",
  call = NULL, package = packageName(), ...)

\code{ FileFormatException( path = NA, line = NA, data = NA,
   message = *, call = NULL, package = packageName(), ...)}

\code{ EmptyException( path = NA, line = NA, data = NA,
   message = *, call = NULL, package = packageName(), ...)}

\code{ FileNotEmptyException( path = NA, line = NA, data = NA,
   message = *, call = NULL, package = packageName(), ...)}
}
\arguments{
\item{data}{The data element associated with an exception. Format varies and
should be described in subclass documentation. This may be \code{NA} but
should not be \code{NULL}. Accessible via the the \code{exceptionData} S3
accessor.}

\item{message}{The message associated with the exception, for reading by
humans. The default message is usually good enough, but can be overridden
through this parameter if needed. When constructed inside packages, the
message displayed will have the name of the package prepended as
\code{'[\var{package}] \var{message}'} due to \code{conditionMessage} being
over-ridden for \code{Exception} subclasses.}

\item{call}{The call or call stack associated with the exception. By default
this is NUL. Can be overridden to provide the \code{\link{Sys.calls}} call
stack or a single \code{\link{Sys.call}}.}

\item{package}{The package where this exception is generated. May be
\code{NULL} when exceptions are generated and signaled from interactive or
script level code and not from a package. Attempts to guess the package by
default. Will usually include the package in the displayed \code{message}.}

\item{...}{Additional arguments defining \code{\var{arg}= \var{value}} data
to include in the Exception (accessible as list elements, but preferably
by S3 accessor functions).}

\item{path}{The path associated with an exception, as a one element character
vector This is required for a \code{FileFormatException} or sub-class. It
may be \code{NA} but should not be \code{NULL}. Accessible via the the
\code{exceptionPath} S3 accessor.}

\item{line}{The line associated with an exception, as a one element integer
vector This is required for a \code{FileFormatException} or sub-class. It
may be \code{NA} but should not be \code{NULL}. Accessible via the the
\code{exceptionLine} S3 accessor.}
}
\value{
An \code{DataException} or descendant describing some problematic
  event involving a data values, names, or formats. Always is/extends
  class \code{c( "DataException", "Exception", "condition" )}.
}
\description{
\code{DataException}s and their sub-classes describe problems with data
values, names, or formats. Subclasses extended the specificity of the
\code{DataException}. Each takes a \code{data} argument and possibly others.
See \code{\link{Exception}}. These exceptions are intended to be signaled,
i.e. with \code{stop} and exist both to provide pre-defined error messages
and so that handlers in try/catch blocks have latitude to choose what
granularity of exception to catch.
}
\section{Functions}{
\itemize{
\item \code{DataException}: This exception and sub-classes are \code{Exception}s that describe
problems with data values, names, or formats.
\itemize{
   \item message: "\code{[\var{package}] A DataException occurred.}"
   \item class: \code{c( "DataException", "Exception", "condition" )}
   \item data: Defaults: \code{data = NA}.
}

\item \code{FileFormatException}: This exception and sub-classes are \code{Exception}s that describe
problems with data files
\itemize{
   \item message: "\code{[\var{package}] A FileFormatException occurred at
       line \var{line} in file: "\var{file}". (Running in: "\var{getwd()}").
       That line began: "\var{first 80 char}".}"
   \item class: \code{c( "FileFormatException", "DataException", "Exception",
      "condition" )}
   \item data: Defaults: \code{path = NA, line = NA, data = NA}.
}

\item \code{EmptyFileException}: This exception indicates a data file was unexpectedly empty
\itemize{
   \item message:  [\var{package}] An EmptyFileException occurred. File
   is unexpectedly empty: "\var{file}". (Running in: "\var{getwd()}").
   \item class: \code{c( "EmptyFileException", "FileFormatException",
       "DataException", "Exception", "condition" )}
   \item data: Defaults: \code{path = NA, data= NA, line= NA}.
}

\item \code{FileNotEmptyException}: This exception indicates a data file that should have been empty but wasn't.
\itemize{
   \item message:  [\var{package}] A FileNotEmptyException occurred. File was
   expectedly to be empty but wasn't: "\var{file}". (Running in:
   "\var{getwd()}").
   \item class: \code{c( "FileNotEmptyException", "FileFormatException",
       "DataException", "Exception", "condition" )}
   \item data: Defaults: \code{path = NA, data= NA, line= NA}.
}
}}
\examples{
dEx <- IOException()
conditionMessage(dEx)
conditionCall(dEx)
exceptionPackage(dEx)
exceptionData(dEx)
}
\seealso{
\code{condition}, \code{conditionMessage()}, \code{conditionCall},
  \code{Exception}, \code{exceptionPackage}, \code{exceptionData}
}

