% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataExceptionsData.R
\name{DataExceptionData}
\alias{DataExceptionData}
\alias{exceptionData}
\alias{exceptionData.Exception}
\alias{exceptionLine}
\alias{exceptionLine.Exception}
\title{Data accessors for DataExceptions}
\usage{
exceptionData(...)

\method{exceptionData}{Exception}(e, ...)

exceptionLine(...)

\method{exceptionLine}{Exception}(e, ...)
}
\arguments{
\item{...}{Required for generic S3 method implementation, not used by any
method defined here.}

\item{e}{The \code{Exception} to query for a data element.}
}
\value{
\describe{
  \item{\code{exceptionData}}{Returns the data causing the exception, format varies).}
  \item{\code{exceptionPath}}{Returns a file system path (as a one-element character vector).}
  \item{\code{exceptionLine}}{Returns a file line number(as a one-element integer vector).}
}
}
\description{
Data elements used in the \code{DataException} hierarchy each have there own S3
based accessor methods which can be used to extract data from an exception if
caught. All accessors can be used on any \code{Exception} derived object,
returning \code{NULL} if the data element is not defined for that exception.
}
\seealso{
\code{\link{DataException}}
}

