% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exceptions.R
\name{exceptionPackage}
\alias{exceptionPackage}
\alias{exceptionPackage.Exception}
\title{Access the package data in an Exception}
\usage{
exceptionPackage(...)

\method{exceptionPackage}{Exception}(e, ...)
}
\arguments{
\item{...}{Future-proof the S3 generic method; not used here.}

\item{e}{The Exception object to query}
}
\value{
The package where the \code{Exception} was generated.
}
\description{
An S3 method to extract the package in which an Exception occurred. Can be
applied to any Exception derived object, but there is no default version.
}
\examples{
e <- Exception()
package <- exceptionPackage(e)

e <- IOException()
package <- exceptionPackage(e)

}
\seealso{
\code{Exception}
}

