% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testUtils.R
\name{expect_condition}
\alias{expect_condition}
\title{Are conditions signaled correctly?}
\usage{
expect_condition(object, class = "condition", label = NULL, info = NULL)
}
\arguments{
\item{object}{Code that is expected to signal a condition.}

\item{class}{Class of the condition signaled, as a character vector. At least
one of these class names must be amongst the condition's classes. By
default this checks for "condition" which is the ubiquitous base class all
signals must inherit.}

\item{label}{The object label. NULL is treated as the string "Condition".}

\item{info}{Extra information to be included in the message}
}
\value{
Called only for its side effects during testing.
}
\description{
Are conditions signaled correctly?
}
\examples{
\dontrun{
### Pass
expect_condition( stop( "Oops" ))   # Checks if condition by default
expect_condition( stop( "Oops" ), "error" )
expect_condition( stop( "Oops" ), "simple-error" )
expect_condition( warning( "Oops" ), "warning" )
expect_condition( warning( "Oops" ), c("Foo", error, "Bar", "simple-error"))

### Fail
# Nothing signaled
expect_condition( "Oops" )

# error, simple-error, or condition, not warning
expect_condition( stop("Oops"), "warning" )

# error, simple-error, or condition, neither Foo nor Bar
expect_condition( stop( "Oops" ), c( "Foo", "Bar" ))
}
}

