% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOExceptionData.R
\name{IOExceptionData}
\alias{IOExceptionData}
\alias{exceptionPath}
\alias{exceptionPath.Exception}
\alias{exceptionTarget}
\alias{exceptionTarget.Exception}
\title{Data accessors for IOExceptions}
\usage{
exceptionPath(...)

\method{exceptionPath}{Exception}(e, ...)

exceptionTarget(...)

\method{exceptionTarget}{Exception}(e, ...)
}
\arguments{
\item{...}{Required for generic S3 method implementation, not used by any
method defined here.}

\item{e}{The \code{Exception} to query for a data element.}
}
\value{
\describe{
  \item{\code{exceptionPath}}{Returns a file system path (as a one-element character vector).}
  \item{\code{exceptionTarget}}{Returns the target of a link (as a one-element character vector).}
}
}
\description{
Data elemets used in the \code{IOException} hierarchy each have there own S3
based accessor methods which can be used to extract data from an exception if
caught. All accessors can be used on any \code{Exception} derived object,
returning \code{NULL} if the data element is not defined for that exception.
}
\seealso{
\code{\link{IOException}}
}

