% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samFilters.R
\name{pairedEndReadTests}
\alias{pairHasRef}
\alias{pairIsAligned}
\alias{pairIsChimeric}
\alias{pairIsNotChimeric}
\alias{pairIsPrimary}
\alias{pairedEndReadTests}
\title{Paired-end read tests}
\usage{
pairIsAligned(read, primary.only = FALSE)

pairIsPrimary(read, primary.only = FALSE)

pairHasRef(read, ref)

pairIsChimeric(read, ref1, ref2)

pairIsNotChimeric(read, ref)
}
\arguments{
\item{read}{The read to test.}

\item{ref}{Regular expression matched against the reference this and/or the
other end aligns too. Keeps a read if \code{ref} matches to either
(\code{rname}) or (\code{rnext}) or both.}

\item{ref1}{Regular expression that matches the chromosome/reference name
of one end, either \code{rname} or \code{rnext}. Should not match both.}

\item{ref2}{Reqular expression that matches the chromosome/reference name of
whichever end \code{ref1} does not match.}
}
\value{
Returns \code{TRUE} if a read passes the test, \code{FALSE}
  otherwise. Each read from the same template will return the same result.
}
\description{
Tests a read to see if its meets some criteria. Only
tests that keep both or delete both reads when applied to each read end
separately are paired-end read tests. These tests are suitable for use in a
\code{\link{samReadFilter}} and will not result in unpaired reads after
application.
}
\section{Functions}{
\itemize{
\item \code{pairIsAligned}: selects paired end reads when both ends
are aligned. Relies on accuracy of READ_UNMAPPED and MATE_UNMAPPED flags and
does not consider any other indicators. This may select multiple reads for one
or both ends.

\item \code{pairIsPrimary}: selects paired end reads when both ends
are aligned and are the primary alignment. Relies on accuracy of the
READ_UNMAPPED, MATE_UNMAPPED, and NOT_PRIMARY_ALIGNMENT flags.

\item \code{pairHasRef}: Identifies paired end reads when one end is aligned
to a selected reference or chromosome. Takes a regular expression and
will return TRUE if either a reads \code{rnext} or \code{rname} (or both)
matches it.

\item \code{pairIsChimeric}: Identifies paired end reads when one end is
aligned to one reference or chromosome and the other end to a different
one. Takes two regular expressions, and will return TRUE only if one read
end matches the first expression and the other read end matches the second.

\item \code{pairIsNotChimeric}: Identifies paired end reads when both ends are
aligned to the same reference or chromosome and the other end to a different
one. Takes a regular expressiona, and will return TRUE only if both a reads
\code{rnext} and \code{rname} match it.
}}
\examples{
\dontrun{
sam <- Sam("someFile")
aRead <- SamReads(sam)[1]

pairHasRef( aRead, "^chrX$|^chrY$" )
sexLinked <- samReadFilter( sam, "pairHasRef", "^chrX$|^chrY$" )

pairIsChimeric(aRead, "^chr..?$", "hiv|HIV")
hivIntegrationSam <- samReadFilter( sam, "pairIsChimeric", "^chr..?$", "hiv|HIV" )

pairIsNotChimeric(aRead, "hiv|HIV")
hivOnly <- samReadFilter( sam, "pairIsNotChimeric", "hiv|HIV" )

}
}

