% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samFilters.R
\name{pairedReads}
\alias{pairedReads}
\title{Filter paired or unpaired reads}
\usage{
pairedReads(sam, paired = TRUE)
}
\arguments{
\item{sam}{The Sam object to filter.}

\item{paired}{Set \code{FALSE} to keep only unpaired reads. Keeps
only paired reads by default.}
}
\value{
A sam object with only paired reads, or with only unpaired reads.
}
\description{
Drop all paired or unpaired reads. Relies on the accuracy and disjoint nature
of the FIRST_OF_PAIR and SECOND_OF_PAIR flags, and that each paired-end read
has a unique id (qname). Maintains multiple alignments for any kept read
ends.
}
\details{
This allows dropping unpaired reads after filtering that might only drop one
read of a pair, such as filtering by quality.
}
\examples{
\dontrun{
sam <- Sam("someFile")
pairedOnly <- pairedReads(sam)
unpairedOnly <- pairedReads(sam, FALSE)
}

}

