% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamHeader.R
\name{parseSamHeaderLines}
\alias{parseSamHeaderLines}
\title{Parse strings as sam file header lines}
\usage{
parseSamHeaderLines(x)
}
\arguments{
\item{x}{Vector of string, each a sam file header}
}
\value{
A data frame with two columns, the first giving the tag name, the
second the original full line. Not the final version of this :)
}
\description{
Parse strings as sam file header lines
}

