% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamReads.R
\name{parseSamReadLines}
\alias{parseSamReadLines}
\title{Parse Sam file read lines}
\usage{
parseSamReadLines(lines, splitTags = TRUE)
}
\arguments{
\item{lines}{A character vector of read lines, in order, from a sam file}

\item{splitTags}{Set \code{FALSE} to leave optional tags unprocessed as a
\code{tags} column of characters (with embedded tabs). By default will split
optional tags out as columns, but this is significantly slower (5x or more).}
}
\value{
A data frame of the parsed read lines, with required column names
  matching the SAM spec. Optional tags, if split out, will each have their
  own column of values named for the two character tag. Columns will be
  integer or double if appropriate, all other types are left as character.
}
\description{
Parse Sam file read lines
}

