% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exceptions.R
\name{path}
\alias{path}
\alias{path.FileException}
\title{Get the path associated with a file system exception.}
\usage{
path(...)

\method{path}{FileException}(e, ...)
}
\arguments{
\item{e}{The file system exception to query}
}
\value{
The path of the thing a file system exception is referring to as a 1
  element character vector.
}
\description{
Get the path associated with a file system exception.
}
\seealso{
FileSystemException, conditionMessage, conditionCall
}

