% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sam.R
\name{sam}
\alias{sam}
\alias{sam<-}
\title{Create a sam (sequence alignment/map) object.}
\usage{
sam(file)

sam(file) <- value
}
\arguments{
\item{file}{A *.sam file name.}

\item{value}{A \code{sam} object to write to a sam file.}
}
\value{
A \code{sam} object. Nothing is returned by the function that writes
a file. Writing to a new file name does not change what is returned by
\code{getFilename(samObj)} or \code{getFileName(samMeta(samObj))}.
}
\description{
An object of class \code{sam} represents a \file{*.sam} file. This is
supposed to be an opaque object for use by other functions. The
size of a \code{sam} object is limited to only what can be read into memory and
manipulated, which is probably at best about 1/3 of the total memory on the
system (due to \R functions being "copy by value".) A \code{sam} object can
be written back to a file by assigning the value to the constructor call.
No support is provided for reading or writing \file{*.bam} files.
}
\section{Internals}{
 A list with three elements:
\itemize{
   \item{"samHeader" - a \code{samHeader} object.}
   \item{"samReads" - a \code{samReads} object.}
   \item{"samMeta" - a \code{samMeta} object.}
}
}
\examples{
mySamFile <- sam( 'myFile.sam' )
sam('newFile.sam') <- mySamFile
tools::md5('myFile.sam') == tools::md5('newFile.sam')

}
\seealso{
\code{samHeader}, \code{samReads}, \code{samMeta}
}

