% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samFilters.R
\name{samReadCount}
\alias{samReadCount}
\title{Count reads meeting filter criteria}
\usage{
samReadCount(sam, FUNC, ..., .progress = "none", .inform = FALSE,
  .parallel = FALSE, .paropts = NULL, .id = NA)
}
\arguments{
\item{sam}{A Sam object}

\item{FUNC}{A function whose first parameter is a single read and that
returns a single \code{TRUE} if that read should be kept when filtering the
\code{sam} object, or \code{FALSE} if not. Should never return \code{NULL}
or \code{NA}.}

\item{...}{Other parameters to pass to \code{FUNC}.}

\item{.progress}{Progress bar; see \code{\link[plyr]{adply}}.}

\item{.inform}{Extra error handling; see \code{\link[plyr]{adply}}.}

\item{.parallel}{Use the parallel package; see \code{\link[plyr]{adply}}.}

\item{.paropts}{Options if using the parallel package; see \code{\link[plyr]{adply}}.}

\item{.id}{Id column; see \code{\link[plyr]{adply}}.}
}
\value{
The count of matching reads
}
\description{
Apply a filter to a sam object and count the reads for which the filter
\code{FUNC} returns \code{TRUE}. Uses \code{\link[plyr]{adply}} to apply this
filter to each read alignment. Many of the parameters are just pass-through
parameters for \code{\link[plyr]{adply}}. These all begin with a "\code{.}"
to distinguish them from potential parameters to the function as they come
after the \code{...} in the function definition. No attempt is made to keep
or delete both ends of a paired end read if one end passes and one end fails
the filter.
}
\examples{
\dontrun{
sam <- Sam("someFile")
hivIntegrationCount <- samReadCount( sam, "readIsOnRef", "^chrMT$" )
}

}
\seealso{
samReadFilter
}

