% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samFilters.R
\name{samReadFilter}
\alias{samReadFilter}
\title{Filter a Sam object}
\usage{
samReadFilter(sam, FUNC, ..., .progress = "none", .inform = FALSE,
  .parallel = FALSE, .paropts = NULL, .id = NA)
}
\arguments{
\item{sam}{A Sam object}

\item{FUNC}{A function whose first parameter is a single read and that
returns a single \code{TRUE} if that read should be kept when filtering the
\code{sam} object, or \code{FALSE} if not. Should never return \code{NULL}
or \code{NA}.}

\item{...}{Other parameters to pass to \code{FUNC}.}

\item{.progress}{Progress bar; see \code{\link[plyr]{adply}}.}

\item{.inform}{Extra error handling; see \code{\link[plyr]{adply}}.}

\item{.parallel}{Use the parallel package; see \code{\link[plyr]{adply}}.}

\item{.paropts}{Options if using the parallel package; see \code{\link[plyr]{adply}}.}

\item{.id}{Id column; see \code{\link[plyr]{adply}}.}
}
\value{
A sam object with only the reads that passed the filter \code{FUNC}.
}
\description{
Apply a filter to a sam object, returing a new sam object with the sam header
but only containing the reads for which the filter \code{FUNC} returns
\code{TRUE}. Uses \code{\link[plyr]{adply}} to apply this filter to each read
alignment. Many of the parameters are just pass-through parameters for
\code{\link[plyr]{adply}}. These all begin with a "\code{.}" to distinguish
them from potential parameters to the function as they come after the
\code{...} in the function definition. No attempt is made to keep or delete
both ends of a paired end read if one end passes and one end fails the
filter.
}
\details{
NOTE: Does not add a @PG header line nor modify any PG:Z tag on any read.
}
\examples{
\dontrun{
sam <- Sam("someFile")
hivIntegrationSam <- samReadFilter( sam, "readIsOnRef", "^chrMT$" )
}

}
\seealso{
Sam, readIsChimeric,
}
