% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-SamSeq.R
\docType{package}
\name{SamSeq}
\alias{SamSeq}
\alias{SamSeq-package}
\title{Work with aligned reads from a SAM file}
\description{
Work with the aligned read data as provided in a *.sam file, as in the format
used by SamTools. Tracks header information, reads, and the source file.
Loads all reads and optionally the extra read tags. Includes functions to
translate the Sam flags and to filter reads.
}
