% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamSource.R
\name{SamSource}
\alias{SamSource}
\alias{SamSource.NULL}
\alias{SamSource.Sam}
\alias{SamSource.character}
\alias{SamSource.list}
\title{Create or retrieve \code{SamSource} objects.}
\usage{
SamSource(x, ...)

\method{SamSource}{character}(x, host = NA, type = NA, ...)

\method{SamSource}{list}(x, ...)

\method{SamSource}{Sam}(x, ...)

\method{SamSource}{NULL}(x, ...)
}
\arguments{
\item{x}{The object to construct a SamSource object from. Can be a Sam,
SamHeader, or SamReads object, a list, or the name of the source file/url
as a character vector, or an explicit NULL. If this is a list ,
\code{"name="}, \code{"host="} and \code{"type="} elements can be specified
as character vectors. Missing or NULL list elements are treated as N/A. All
character vectors should be the same length. If this is an explicit
\code{NULL}, it is treated the same as if it was an empty list.}

\item{...}{Required for S3 object method implementation. Not currently used.}

\item{host}{The name of the system the sam file is available on. Default is
\code{NA}.}

\item{type}{The type of source for the sam file. Currently only "file" is
supported. "url" may be added. The default is NA.}
}
\value{
Returns a SamSource object, an opaque object from which the source
  file name, host machine, and type of access can be extracted.
}
\description{
A \code{SamSource} object is meant to be an opaque object used to store the
source information associated with a Sam file. Currently only \code{type=
"file"} is supported, but \code{"url"} may be added. The name, host machine
name, and type can be retrieved from \code{SamSource} or \code{Sam} objects
with \code{samSourceName()}, \code{samSourceHost()}, and
\code{samSourceType()}, respectively. A \code{SamSource} object can be
retrieved from a \code{Sam}, \code{SamHeader}, or \code{SamReads} object
using \code{SamSource()}.
}
\examples{
\dontrun{
  sam <- Sam("file.sam")
  stopIfNot( SamSource( sam ) == SamSource( SamHeader( sam ) ==
  	SamSource( SamReads( sam )))

  source2 <- SamSource( "file.sam", host="bioinf.unc.edu", type="file")

  sourceList <- list(name="file.sam", host="bioinf.unc.edu", type="file")
  source3 <- SamSource( sourceList )

  source4 <- SamSource(NULL)
}
}

