% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamSource.R
\name{samSourceHost}
\alias{samSourceHost}
\alias{samSourceHost.Sam}
\alias{samSourceHost.SamSource}
\title{SamSource host data accessor}
\usage{
samSourceHost(x, ...)

\method{samSourceHost}{SamSource}(x, ...)

\method{samSourceHost}{Sam}(x, ...)
}
\arguments{
\item{x}{Can be a Sam or SamSource object.}
}
\value{
The name name of the host machine where the sam file the data was
  taken from resides as a character vector.
}
\description{
Accessor to extract the name of the host machine where the sam file the data
was taken from resides.
}

