% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamSource.R
\name{samSourceName}
\alias{samSourceName}
\title{SamSource name data accessor}
\usage{
samSourceName(x)
}
\arguments{
\item{x}{Any object for which \code{SamSource(x)} is defined.}
}
\value{
The name of the source sam data file as a character vector.
}
\description{
Accessor to extract the name of the sam file data was taken from.
}
