% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamSource.R
\name{samSourceType}
\alias{samSourceType}
\alias{samSourceType.Sam}
\alias{samSourceType.SamSource}
\title{SamSource data type accessor}
\usage{
samSourceType(x, ...)

\method{samSourceType}{SamSource}(x, ...)

\method{samSourceType}{Sam}(x, ...)
}
\arguments{
\item{x}{Can be a Sam or SamSource object.}

\item{...}{Required for S3 object method implementation. Not currently used.}
}
\value{
The type of access required for the sam file the data was taken from,
  as a character vector. Can only be "file" currently. "url" is a likely
  future extension.
}
\description{
Accessor to extract the type of access required for the sam file the data was
taken from resides.
}

