% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samFlags.R
\name{testSamFlags}
\alias{allSetSamFlags}
\alias{allUnsetSamFlags}
\alias{anySetSamFlags}
\alias{anyUnsetSamFlags}
\alias{testSamFlags}
\title{Test if sam flags are set or unset.}
\usage{
allSetSamFlags(val, flags)

allUnsetSamFlags(val, flags)

anySetSamFlags(val, flags)

anyUnsetSamFlags(val, flags)
}
\arguments{
\item{val}{The encoded sam flag value to check, as a single integer value}

\item{flags}{The flags to test, as a character vector.}
}
\value{
A single TRUE or FALSE value
}
\description{
Logically compares one or more sam flags against an encoded sam value.
}
\details{
Note: \code{allSetSamFlags} == \code{! anyUnsetSamFlags}, and
\code{allUnsetSamFlags} == \code{! anySetSamFlags}
}
\section{Functions}{
\itemize{
\item \code{allSetSamFlags}: TRUE only if all specified flags are set, FALSE if
any specified flag is unset. The status of unspecified flags does not
matter.

\item \code{allUnsetSamFlags}: TRUE only if all specified flags are unset, FALSE if
any specified flags is set. The status of unspecified flags does not
matter.

\item \code{anySetSamFlags}: TRUE if any specified flag is set, FALSE only if all
specified flags are unset. The status of unspecified flags does not matter.

\item \code{anyUnsetSamFlags}: TRUE if any specified flag is unset, FALSE only if
all specified flags are set. The status of unspecified flags does not
matter.
}}
\examples{
# READ_PAIRED = 1, PROPER_PAIR = 2
# All TRUE
allSetSamFlags( 1 + 2, c( "READ_PAIRED", "PROPER_PAIR" ))
anySetSamFlags( 1 + 2, c( "READ_PAIRED", "PROPER_PAIR" ))
anySetSamFlags( 1 + 2, "READ_PAIRED")
allUnsetSamFlags( 4, c( "READ_PAIRED", "PROPER_PAIR" ))
anyUnsetSamFlags( 1, "PROPER_PAIR" ))
anyUnsetSamFlags( 1, c( "READ_PAIRED", "PROPER_PAIR" ))

# All FALSE
allSetSamFlags( 1, "PROPER_PAIR" ))
allSetSamFlags( 1, c( "READ_PAIRED", "PROPER_PAIR" ))
anySetSamFlags( 4, c( "READ_PAIRED", "PROPER_PAIR" ))
allUnsetSamFlags( 1 + 2, c( "READ_PAIRED", "PROPER_PAIR" ))
allUnsetSamFlags( 1 + 2, "READ_PAIRED" )
anyUnsetSamFlags( 1 + 2, c( "READ_PAIRED", "PROPER_PAIR" ))

}

