% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamReads.R
\name{parseSamReadTags}
\alias{parseSamReadTags}
\title{Parse the optional sam read tags into a data frame}
\usage{
parseSamReadTags(tagList, startAt = 12)
}
\arguments{
\item{tagList}{An un-named list with one element per read line, in order, where each
element is an un-named list, one element per field, where each field after a
given column (by default 12) is a key:typeFlag:value optional sam tag. Order
of the tags in the list is unimportant (although it does affect the order
of the tags in the returned data frame).}

\item{startAt}{The first column in the field list that is to be parsed. All
columns after this are assumed to be optional tag fields. By default the
12'th column is the start of the optional tags.}
}
\value{
A data frame with one column per two character parsed tag. Values are
converted to integer if typeFlag is i and to double if typeFlag is f.
Otherwise left as character. Any tag that is found only on some lines will
have a value of NA on any line where it is not found.
}
\description{
Parse the optional sam read tags into a data frame
}

