% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamHeader.R
\name{SamHeader}
\alias{SamHeader}
\title{Return a SamHeader object}
\usage{
SamHeader(x, ...)
}
\arguments{
\item{x}{The object to extract or build a sam header object from. x can be a
character vector of unparsed header lines, or a Sam object.}

\item{source}{The SamSource object to associate with a header. If
\code{x} is a Sam object, it will use that objects SamSource, otherwise it
will use a default \code{SamSource()}}
}
\description{
S3 method that returns an S3 SamHeader objet.
}

