% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamHeader.R
\name{SamHeader}
\alias{SamHeader}
\alias{SamHeader.character}
\alias{SamHeader.Sam}
\alias{as.data.frame.SamHeader}
\title{Return a SamHeader object}
\usage{
SamHeader(x, ...)

\method{SamHeader}{character}(x, source = SamSource(NULL), ...)

\method{SamHeader}{Sam}(x, ...)

\method{as.data.frame}{SamHeader}(x, ...)
}
\arguments{
\item{x}{For \code{SamHeader(x)}, this is the object to extract or build a
sam header object from. It can be a character vector of unparsed header
lines or a \code{Sam} object. For \code{as.data.frame(x)}, it can only be a
\code{SamHeader} object.}

\item{...}{Required for S3 object method implementation. Not currently used.}

\item{source}{The \code{SamSource} object to associate with a header. If
\code{x} is a \code{Sam} object, it will use that objects \code{SamSource},
otherwise it will use a default \code{SamSource()}}
}
\value{
For \code{SamHeader()}, a \code{SamHeader} object.

  For \code{as.data.frame(SamHeader)}, a data frame as generated by
  \code{\link{parseSamHeaderLines}},
}
\description{
S3 method that returns an S3 SamHeader objet.
}
\section{Internal implementation}{

  A \code{SamHeader} object is supposed to be
  opaque to the user, with internals accesible via \code{SamSource(x)},
  \code{as.data.frame(x)}, and \code{areHeaderTagsParsed(x)}.

  Currently this object is structured as a \code{list} with two elements:
  \code{header} which is the header data frame and
  \code{areHeaderTagsParsed}, which is \code{TRUE} when the record sub-tags
  are parsed parsed and \code{FALSE} otherwise. Currently this is always
  false. The object also has an attribute, \code{source}, which is the source
  object. This structure may change on a patch release with no notice.
}

\seealso{
Sam, SamSource, as.data.frame.SamSource
}
