% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamReads.R
\name{SamReads}
\alias{SamReads}
\alias{SamReads.character}
\alias{SamReads.Sam}
\alias{as.data.frame.SamReads}
\title{Return a SamReads object}
\usage{
SamReads(x, ...)

\method{SamReads}{character}(x, source = SamSource(NULL),
  splitTags = TRUE, ...)

\method{SamReads}{Sam}(x, ...)

\method{as.data.frame}{SamReads}(x, ...)
}
\arguments{
\item{x}{For the object to extract or build a sam reads object from. It can
be a character vector of unparsed read lines, a \code{Sam} object, or
(pretty uselessly) a \code{SamReads} object.}

\item{...}{Future-proof the method. Not currently used.}

\item{source}{The \code{SamSource} object to associate with the reads. If
not provided and \code{x} is a \code{Sam} or \code{SamReads} object, it
will default to that object's source. Otherwise the default is all missing
data, i.e. \code{SamSource(NULL)}}

\item{splitTags}{By default this is \code{TRUE}, meaning optional read tags
are split out as named columns. Setting this \code{FALSE} leaves the
optional tags as a single column, "tags", with all the extra tags
presented as a single tab delimited string.}
}
\value{
A \code{SamReads} object.
}
\description{
S3 method that returns an S3 SamReads object, which is a data frame with
additional meta data.
}
