% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamReads.R
\name{SamReads}
\alias{SamReads}
\alias{SamReads.Sam}
\alias{SamReads.character}
\title{Return a SamReads object}
\usage{
SamReads(x, ...)

\method{SamReads}{character}(x, source = SamSource(NULL), splitTags = TRUE,
  ...)

\method{SamReads}{Sam}(x, ...)
}
\arguments{
\item{x}{The object to extract or build a sam reads object from. \code{x} can be a
character vector of unparsed read lines, or a \code{Sam} object.}

\item{samSource}{The SamSource object to associate with the reads. If not
provided and \code{x} is a \code{Sam} object, it will just use its source.
For other methods, the default is all missing data, i.e.
\code{SamSource(NULL)}}
}
\value{
A SamReads object, which is a data frame.
}
\description{
S3 method that returns an S3 SamReads object, which is a data frame with
additional meta data.
}

