% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamReads.R
\name{SamReads}
\alias{SamReads}
\alias{SamReads.Sam}
\alias{SamReads.character}
\title{Return a SamReads object}
\usage{
SamReads(x, ...)

\method{SamReads}{character}(x, source = SamSource(NULL), splitTags = TRUE,
  ...)

\method{SamReads}{Sam}(x, ...)
}
\arguments{
\item{x}{The object to extract or build a sam reads object from. \code{x} can
be a character vector of unparsed read lines, or a \code{Sam} object.}

\item{...}{Required for S3 object method implementation. Not currently used.}

\item{source}{The SamSource object to associate with the reads. If not
provided and \code{x} is a \code{Sam} object, it will just use its source.
For other methods, the default is all missing data, i.e.
\code{SamSource(NULL)}}

\item{splitTags}{By default this is \code{TRUE}, meaning optional read tags are
split out as named columns. Setting this \code{FALSE} leaves the optional tags as a
single column, \code{tags}, with all optional tags presented as a single tab
delimited string.}
}
\value{
A SamReads object, which can be used as a data frame.
}
\description{
S3 method that returns an S3 SamReads object, which is a data frame with
additional meta data.
}

