% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamSource.R
\name{samSourceHost}
\alias{samSourceHost}
\title{SamSource host data accessor}
\usage{
samSourceHost(x)
}
\arguments{
\item{x}{Any object for which \code{SamSource(x)} is defined.}
}
\value{
The name name of the host machine where the sam file the data was
  taken from resides as a character vector.
}
\description{
Accessor to extract the name of the host machine where the sam file the data
was taken from resides.
}
