% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamSource.R
\name{samSourceName}
\alias{samSourceName}
\alias{samSourceName.Sam}
\alias{samSourceName.SamSource}
\title{SamSource name data accessor}
\usage{
samSourceName(x, ...)

\method{samSourceName}{SamSource}(x, ...)

\method{samSourceName}{Sam}(x, ...)
}
\arguments{
\item{x}{Can be a Sam or SamSource object.}

\item{...}{Required for S3 object method implementation. Not currently used.}
}
\value{
The name of the source sam data file as a character vector.
}
\description{
Accessor to extract the name of the sam file data was taken from.
}

