% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamSource.R
\name{samSourceType}
\alias{samSourceType}
\title{SamSource data type accessor}
\usage{
samSourceType(x)
}
\arguments{
\item{x}{Any object for which \code{SamSource(x)} is defined.}
}
\value{
The type of access required for the sam file the data was taken from,
  as a character vector. Can only be "file" currently. "url" is a likely
  future extension.
}
\description{
Accessor to extract the type of access required for the sam file the data was
taken from resides.
}
